Move LLM CLI to popgetter_cli
