+++ Cargo.toml
@@ -43,3 +43,5 @@ wkt = "0.10.3"
 indoc = "2.0.5"
 langchain-rust = "4.6.0"
 rand = "0.8.5"
+qdrant-client = "1.12.1"
+tiktoken-rs = "0.6.0"
