--- popgetter_cli/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod cli;
-pub mod display;
